+++ src/textual/widgets/__init__.py
@@ -27,6 +27,7 @@ if typing.TYPE_CHECKING:
     from ._loading_indicator import LoadingIndicator
     from ._log import Log
     from ._markdown import Markdown, MarkdownViewer
+    from ._masked_input import MaskedInput
     from ._option_list import OptionList
     from ._placeholder import Placeholder
     from ._pretty import Pretty
@@ -66,6 +67,7 @@ __all__ = [
     "Log",
     "Markdown",
     "MarkdownViewer",
+    "MaskedInput",
     "OptionList",
     "Placeholder",
     "Pretty",